An interface to eight compartmental epidemic models.

epi() is an application programming interface (for the Shiny application that motivated its creation) or a convenient function for interactive use in epidemic modelling. The arguments provided to the function determine compartments used in the model that is run.

Usage

epi(
  population,
  susceptible,
  exposed = 0,
  infected,
  recovered,
  dead = 0,
  beta,
  gamma,
  sigma = 0,
  delta = 0,
  xi = 0,
  muB = 0,
  muD = 0,
  vitalDynamics = FALSE,
  trueMassAction = FALSE,
  timesteps = 25,
  timeIncrement = 1
)

Arguments

population

The initial number of persons in the simulation; if vitalDynamics = TRUE and muB and muD are equal, then it is constant.

susceptible

The initial number of people who are susceptible to the disease (they have no natural immunity to it).

exposed

The initial number of people who have been exposed to the infection and may become infectious themselves.

infected

The initial number of people who are infectious and may infect others.

recovered

The initial number of people whom are no longer infectious and not a risk to others.

dead

The initial number of fatalities (people killed by the disease).

beta

The rate of becoming infectious (if sigma = 0, the default), or the rate of susceptible people being exposed to the infectious.

gamma

The rate of becoming recovered (no longer infectious, if sigma = 0, the default), or the rate of becoming infectious after exposure.

sigma

The rate of becoming recovered (no longer infectious).

delta

The rate of fatality.

xi

The rate of loss of learned immunity.

muB

The rate of birth.

muD

The rate of death not due to the disease under consideration.

vitalDynamics

Whether vital dynamics is enabled (if FALSE, the default, implies muB and muD of zero).

trueMassAction

Whether the force of infection is scaled by the population or not.

timesteps

The number of output rows in the simulation.

timeIncrement

The number of increments of "time" that pass between each output row.

Value

A dataframe with a column for each enabled compartment and a column for time, with timesteps rows. Each row contains the values of the differential equations for each point in time of the model (per function arguments).

Details

The convention for pronouncing epi() is "epi epi" as if you were saying "epidemic" twice. If you do want to refer to the package alone, and not the singular function it packages, you can say "eh pee eye", like "API".

"eh pee eye" is a tongue-in-cheek Canadianism. You might understand it as, "Eh, P.I.?", or just a Canadian slant on API. The latter was thought of first.

While ehpi provides "an API" for the Episim Shiny application, it won't enable the development of applications that provide any feature episim doesn't already provide by wrapping this package. The function is packages was extracted from that project to allow associating well-formatted documentation, vignettes, and examples with the function because that function is expected to be used outside of Episim, interactively, during workshops run by Dr. Krishnamurthy.

Examples

epi(population = 500,
    susceptible = 489,
    exposed = 10,
    infected = 1,
    recovered = 0,
    dead = 29,
    beta = 0.005,
    gamma = 0.0009,
    sigma = 0.000001,
    delta = 0.01,
    timesteps = 230)
#>     time        N            S         E         I            R        D
#> 1      1 500.0000 4.890000e+02  10.00000  1.000000 0.000000e+00 29.00000
#> 2      2 499.9900 4.865614e+02  12.42847  1.000094 9.998648e-07 29.01000
#> 3      3 499.9800 4.841322e+02  14.84546  1.002357 2.000908e-06 29.02001
#> 4      4 499.9699 4.817070e+02  17.25618  1.006758 3.005288e-06 29.03005
#> 5      5 499.9598 4.792809e+02  19.66567  1.013273 4.015128e-06 29.04015
#> 6      6 499.9497 4.768490e+02  22.07881  1.021883 5.032533e-06 29.05033
#> 7      7 499.9394 4.744065e+02  24.50032  1.032571 6.059588e-06 29.06060
#> 8      8 499.9290 4.719489e+02  26.93480  1.045327 7.098365e-06 29.07098
#> 9      9 499.9185 4.694716e+02  29.38672  1.060143 8.150928e-06 29.08151
#> 10    10 499.9078 4.669704e+02  31.86041  1.077018 9.219338e-06 29.09219
#> 11    11 499.8969 4.644409e+02  34.36012  1.095951 1.030565e-05 29.10306
#> 12    12 499.8859 4.618790e+02  36.88996  1.116947 1.141193e-05 29.11412
#> 13    13 499.8746 4.592806e+02  39.45398  1.140015 1.254023e-05 29.12540
#> 14    14 499.8631 4.566418e+02  42.05611  1.165166 1.369265e-05 29.13693
#> 15    15 499.8513 4.539587e+02  44.70018  1.192416 1.487127e-05 29.14871
#> 16    16 499.8392 4.512275e+02  47.38996  1.221782 1.607819e-05 29.16078
#> 17    17 499.8268 4.484444e+02  50.12911  1.253287 1.731554e-05 29.17316
#> 18    18 499.8141 4.456060e+02  52.92120  1.286955 1.858548e-05 29.18585
#> 19    19 499.8011 4.427085e+02  55.76975  1.322813 1.989018e-05 29.19890
#> 20    20 499.7877 4.397486e+02  58.67815  1.360892 2.123185e-05 29.21232
#> 21    21 499.7739 4.367229e+02  61.64973  1.401225 2.261272e-05 29.22613
#> 22    22 499.7596 4.336281e+02  64.68771  1.443847 2.403506e-05 29.24035
#> 23    23 499.7450 4.304609e+02  67.79524  1.488796 2.550119e-05 29.25501
#> 24    24 499.7299 4.272184e+02  70.97536  1.536113 2.701344e-05 29.27013
#> 25    25 499.7143 4.238974e+02  74.23101  1.585841 2.857422e-05 29.28574
#> 26    26 499.6981 4.204951e+02  77.56501  1.638024 3.018594e-05 29.30186
#> 27    27 499.6815 4.170087e+02  80.98009  1.692710 3.185110e-05 29.31851
#> 28    28 499.6643 4.134355e+02  84.47884  1.749948 3.357222e-05 29.33572
#> 29    29 499.6465 4.097729e+02  88.06373  1.809787 3.535186e-05 29.35352
#> 30    30 499.6281 4.060187e+02  91.73709  1.872281 3.719267e-05 29.37193
#> 31    31 499.6090 4.021704e+02  95.50111  1.937482 3.909733e-05 29.39097
#> 32    32 499.5893 3.982260e+02  99.35782  2.005448 4.106856e-05 29.41069
#> 33    33 499.5689 3.941836e+02 103.30908  2.076233 4.310916e-05 29.43109
#> 34    34 499.5478 3.900413e+02 107.35656  2.149894 4.522198e-05 29.45222
#> 35    35 499.5259 3.857976e+02 111.50178  2.226492 4.740993e-05 29.47410
#> 36    36 499.5032 3.814511e+02 115.74601  2.306083 4.967597e-05 29.49676
#> 37    37 499.4798 3.770007e+02 120.09033  2.388728 5.202311e-05 29.52023
#> 38    38 499.4555 3.724453e+02 124.53561  2.474486 5.445446e-05 29.54454
#> 39    39 499.4303 3.677844e+02 129.08244  2.563417 5.697314e-05 29.56973
#> 40    40 499.4042 3.630174e+02 133.73118  2.655581 5.958237e-05 29.59582
#> 41    41 499.3771 3.581441e+02 138.48193  2.751036 6.228540e-05 29.62285
#> 42    42 499.3491 3.531648e+02 143.33448  2.849842 6.508556e-05 29.65086
#> 43    43 499.3201 3.480797e+02 148.28835  2.952055 6.798622e-05 29.67986
#> 44    44 499.2901 3.428895e+02 153.34274  3.057732 7.099082e-05 29.70991
#> 45    45 499.2590 3.375954e+02 158.49655  3.166929 7.410286e-05 29.74103
#> 46    46 499.2267 3.321986e+02 163.74832  3.279699 7.732587e-05 29.77326
#> 47    47 499.1934 3.267009e+02 169.09626  3.396092 8.066347e-05 29.80663
#> 48    48 499.1588 3.211043e+02 174.53824  3.516159 8.411928e-05 29.84119
#> 49    49 499.1230 3.154113e+02 180.07174  3.639946 8.769702e-05 29.87697
#> 50    50 499.0860 3.096245e+02 185.69388  3.767496 9.140043e-05 29.91400
#> 51    51 499.0477 3.037473e+02 191.40143  3.898850 9.523328e-05 29.95233
#> 52    52 499.0080 2.977831e+02 197.19074  4.034046 9.919941e-05 29.99199
#> 53    53 498.9670 2.917360e+02 203.05779  4.173115 1.033027e-04 30.03303
#> 54    54 498.9245 2.856102e+02 208.99818  4.316088 1.075469e-04 30.07547
#> 55    55 498.8806 2.794104e+02 215.00712  4.462989 1.119362e-04 30.11936
#> 56    56 498.8353 2.731419e+02 221.07944  4.613838 1.164742e-04 30.16474
#> 57    57 498.7883 2.668100e+02 227.20959  4.768650 1.211651e-04 30.21165
#> 58    58 498.7399 2.604206e+02 233.39166  4.927435 1.260129e-04 30.26013
#> 59    59 498.6898 2.539801e+02 239.61940  5.090197 1.310213e-04 30.31021
#> 60    60 498.6381 2.474948e+02 245.88618  5.256934 1.361946e-04 30.36195
#> 61    61 498.5846 2.409718e+02 252.18508  5.427639 1.415365e-04 30.41537
#> 62    62 498.5295 2.344182e+02 258.50886  5.602298 1.470512e-04 30.47051
#> 63    63 498.4726 2.278415e+02 264.85002  5.780890 1.527424e-04 30.52742
#> 64    64 498.4139 2.212495e+02 271.20077  5.963389 1.586143e-04 30.58614
#> 65    65 498.3533 2.146502e+02 277.55313  6.149759 1.646705e-04 30.64671
#> 66    66 498.2908 2.080518e+02 283.89889  6.339962 1.709151e-04 30.70915
#> 67    67 498.2265 2.014627e+02 290.22971  6.533948 1.773517e-04 30.77352
#> 68    68 498.1602 1.948912e+02 296.53710  6.731664 1.839842e-04 30.83984
#> 69    69 498.0918 1.883461e+02 302.81249  6.933047 1.908162e-04 30.90816
#> 70    70 498.0215 1.818360e+02 309.04728  7.138027 1.978515e-04 30.97851
#> 71    71 497.9491 1.753695e+02 315.23284  7.346530 2.050935e-04 31.05093
#> 72    72 497.8745 1.689553e+02 321.36059  7.558472 2.125457e-04 31.12546
#> 73    73 497.7979 1.626019e+02 327.42201  7.773764 2.202115e-04 31.20212
#> 74    74 497.7191 1.563178e+02 333.40875  7.992308 2.280943e-04 31.28094
#> 75    75 497.6380 1.501112e+02 339.31257  8.214002 2.361972e-04 31.36197
#> 76    76 497.5548 1.439903e+02 345.12550  8.438737 2.445233e-04 31.44523
#> 77    77 497.4692 1.379628e+02 350.83977  8.666397 2.530757e-04 31.53076
#> 78    78 497.3814 1.320363e+02 356.44796  8.896862 2.618571e-04 31.61857
#> 79    79 497.2913 1.262181e+02 361.94295  9.130005 2.708703e-04 31.70870
#> 80    80 497.1988 1.205148e+02 367.31800  9.365696 2.801179e-04 31.80118
#> 81    81 497.1040 1.149331e+02 372.56679  9.603799 2.896025e-04 31.89602
#> 82    82 497.0067 1.094788e+02 377.68344  9.844174 2.993263e-04 31.99326
#> 83    83 496.9071 1.041576e+02 382.66253 10.086678 3.092915e-04 32.09292
#> 84    84 496.8050 9.897438e+01 387.49913 10.331164 3.195003e-04 32.19500
#> 85    85 496.7005 9.393379e+01 392.18885 10.577482 3.299545e-04 32.29954
#> 86    86 496.5934 8.903980e+01 396.72782 10.825482 3.406558e-04 32.40656
#> 87    87 496.4839 8.429587e+01 401.11271 11.075010 3.516059e-04 32.51606
#> 88    88 496.3719 7.970490e+01 405.34076 11.325911 3.628063e-04 32.62806
#> 89    89 496.2574 7.526925e+01 409.40976 11.578030 3.742582e-04 32.74258
#> 90    90 496.1404 7.099071e+01 413.31807 11.831213 3.859627e-04 32.85963
#> 91    91 496.0208 6.687051e+01 417.06458 12.085303 3.979209e-04 32.97921
#> 92    92 495.8987 6.290934e+01 420.64876 12.340148 4.101336e-04 33.10134
#> 93    93 495.7740 5.910737e+01 424.07060 12.595593 4.226014e-04 33.22601
#> 94    94 495.6468 5.546421e+01 427.33061 12.851488 4.353249e-04 33.35325
#> 95    95 495.5170 5.197902e+01 430.42980 13.107683 4.483045e-04 33.48304
#> 96    96 495.3846 4.865044e+01 433.36966 13.364033 4.615403e-04 33.61540
#> 97    97 495.2497 4.547669e+01 436.15212 13.620394 4.750325e-04 33.75033
#> 98    98 495.1122 4.245553e+01 438.77954 13.876626 4.887811e-04 33.88781
#> 99    99 494.9721 3.958436e+01 441.25469 14.132592 5.027857e-04 34.02786
#> 100  100 494.8295 3.686018e+01 443.58068 14.388161 5.170461e-04 34.17046
#> 101  101 494.6844 3.427970e+01 445.76095 14.643203 5.315618e-04 34.31562
#> 102  102 494.5367 3.183929e+01 447.79925 14.897597 5.463323e-04 34.46332
#> 103  103 494.3864 2.953508e+01 449.69957 15.151221 5.613568e-04 34.61357
#> 104  104 494.2337 2.736296e+01 451.46616 15.403964 5.766345e-04 34.76634
#> 105  105 494.0784 2.531863e+01 453.10342 15.655715 5.921644e-04 34.92164
#> 106  106 493.9205 2.339763e+01 454.61593 15.906371 6.079455e-04 35.07946
#> 107  107 493.7602 2.159537e+01 456.00841 16.155832 6.239767e-04 35.23977
#> 108  108 493.5974 1.990715e+01 457.28564 16.404007 6.402568e-04 35.40257
#> 109  109 493.4322 1.832820e+01 458.45249 16.650805 6.567843e-04 35.56784
#> 110  110 493.2644 1.685375e+01 459.51385 16.896145 6.735579e-04 35.73558
#> 111  111 493.0942 1.547897e+01 460.47463 17.139948 6.905761e-04 35.90576
#> 112  112 492.9216 1.419907e+01 461.33971 17.382143 7.078372e-04 36.07837
#> 113  113 492.7466 1.300929e+01 462.11393 17.622660 7.253398e-04 36.25340
#> 114  114 492.5692 1.190493e+01 462.80207 17.861439 7.430820e-04 36.43082
#> 115  115 492.3894 1.088137e+01 463.40883 18.098421 7.610621e-04 36.61062
#> 116  116 492.2072 9.934082e+00 463.93880 18.333554 7.792782e-04 36.79278
#> 117  117 492.0227 9.058648e+00 464.39648 18.566788 7.977285e-04 36.97729
#> 118  118 491.8359 8.250773e+00 464.78622 18.798080 8.164111e-04 37.16411
#> 119  119 491.6468 7.506297e+00 465.11224 19.027391 8.353240e-04 37.35324
#> 120  120 491.4553 6.821206e+00 465.37860 19.254685 8.544652e-04 37.54465
#> 121  121 491.2617 6.191633e+00 465.58923 19.479931 8.738327e-04 37.73833
#> 122  122 491.0658 5.613871e+00 465.74789 19.703100 8.934244e-04 37.93424
#> 123  123 490.8676 5.084371e+00 465.85817 19.924168 9.132382e-04 38.13238
#> 124  124 490.6673 4.599752e+00 465.92348 20.143115 9.332720e-04 38.33272
#> 125  125 490.4648 4.156793e+00 465.94709 20.359922 9.535237e-04 38.53524
#> 126  126 490.2601 3.752441e+00 465.93210 20.574576 9.739912e-04 38.73991
#> 127  127 490.0533 3.383808e+00 465.88141 20.787064 9.946722e-04 38.94672
#> 128  128 489.8444 3.048165e+00 465.79780 20.997377 1.015565e-03 39.15565
#> 129  129 489.6333 2.742943e+00 465.68385 21.205509 1.036666e-03 39.36666
#> 130  130 489.4203 2.465731e+00 465.54201 21.411454 1.057975e-03 39.57975
#> 131  131 489.2051 2.214265e+00 465.37456 21.615212 1.079488e-03 39.79488
#> 132  132 488.9880 1.986431e+00 465.18364 21.816781 1.101205e-03 40.01205
#> 133  133 488.7688 1.780255e+00 464.97125 22.016164 1.123121e-03 40.23121
#> 134  134 488.5476 1.593897e+00 464.73923 22.213364 1.145236e-03 40.45236
#> 135  135 488.3245 1.425648e+00 464.48933 22.408385 1.167547e-03 40.67547
#> 136  136 488.0995 1.273923e+00 464.22313 22.601234 1.190052e-03 40.90052
#> 137  137 487.8725 1.137255e+00 463.94212 22.791920 1.212749e-03 41.12749
#> 138  138 487.6436 1.014287e+00 463.64767 22.980449 1.235635e-03 41.35635
#> 139  139 487.4129 9.037675e-01 463.34105 23.166834 1.258709e-03 41.58709
#> 140  140 487.1803 8.045446e-01 463.02341 23.351084 1.281968e-03 41.81968
#> 141  141 486.9459 7.155595e-01 462.69582 23.533212 1.305411e-03 42.05411
#> 142  142 486.7097 6.358405e-01 462.35926 23.713230 1.329034e-03 42.29034
#> 143  143 486.4716 5.644975e-01 462.01464 23.891151 1.352836e-03 42.52836
#> 144  144 486.2318 5.007164e-01 461.66276 24.066990 1.376816e-03 42.76816
#> 145  145 485.9903 4.437536e-01 461.30439 24.240761 1.400970e-03 43.00970
#> 146  146 485.7470 3.929316e-01 460.94020 24.412479 1.425296e-03 43.25296
#> 147  147 485.5021 3.476333e-01 460.57082 24.582160 1.449794e-03 43.49794
#> 148  148 485.2554 3.072978e-01 460.19681 24.749820 1.474460e-03 43.74460
#> 149  149 485.0071 2.714162e-01 459.81868 24.915474 1.499293e-03 43.99293
#> 150  150 484.7571 2.395270e-01 459.43691 25.079139 1.524290e-03 44.24290
#> 151  151 484.5055 2.112126e-01 459.05190 25.240833 1.549451e-03 44.49451
#> 152  152 484.2523 1.860957e-01 458.66404 25.400571 1.574771e-03 44.74771
#> 153  153 483.9975 1.638355e-01 458.27368 25.558372 1.600251e-03 45.00251
#> 154  154 483.7411 1.441249e-01 457.88112 25.714251 1.625887e-03 45.25887
#> 155  155 483.4832 1.266875e-01 457.48664 25.868227 1.651679e-03 45.51679
#> 156  156 483.2238 1.112746e-01 457.09050 26.020317 1.677623e-03 45.77623
#> 157  157 482.9628 9.766308e-02 456.69291 26.170537 1.703719e-03 46.03719
#> 158  158 482.7004 8.565257e-02 456.29407 26.318907 1.729964e-03 46.29964
#> 159  159 482.4364 7.506374e-02 455.89418 26.465442 1.756356e-03 46.56356
#> 160  160 482.1711 6.573608e-02 455.49338 26.610160 1.782894e-03 46.82894
#> 161  161 481.9042 5.752612e-02 455.09183 26.753079 1.809576e-03 47.09576
#> 162  162 481.6360 5.030579e-02 454.68965 26.894216 1.836400e-03 47.36400
#> 163  163 481.3664 4.396087e-02 454.28695 27.033588 1.863364e-03 47.63364
#> 164  164 481.0953 3.838963e-02 453.88385 27.171213 1.890466e-03 47.90466
#> 165  165 480.8229 3.350152e-02 453.48042 27.307106 1.917706e-03 48.17706
#> 166  166 480.5492 2.921608e-02 453.07675 27.441286 1.945080e-03 48.45080
#> 167  167 480.2741 2.546184e-02 452.67292 27.573769 1.972588e-03 48.72588
#> 168  168 479.9977 2.217542e-02 452.26898 27.704572 2.000227e-03 49.00227
#> 169  169 479.7200 1.930064e-02 451.86500 27.833712 2.027996e-03 49.27996
#> 170  170 479.4411 1.678776e-02 451.46101 27.961205 2.055894e-03 49.55894
#> 171  171 479.1608 1.459281e-02 451.05708 28.087067 2.083918e-03 49.83918
#> 172  172 478.8793 1.267691e-02 450.65322 28.211316 2.112067e-03 50.12067
#> 173  173 478.5966 1.100576e-02 450.24949 28.333966 2.140340e-03 50.40340
#> 174  174 478.3127 9.549086e-03 449.84590 28.455034 2.168735e-03 50.68735
#> 175  175 478.0275 8.280232e-03 449.44249 28.574536 2.197250e-03 50.97250
#> 176  176 477.7412 7.175718e-03 449.03928 28.692488 2.225883e-03 51.25883
#> 177  177 477.4537 6.214895e-03 448.63628 28.808905 2.254634e-03 51.54634
#> 178  178 477.1650 5.379613e-03 448.23353 28.923803 2.283501e-03 51.83501
#> 179  179 476.8752 4.653937e-03 447.83103 29.037196 2.312481e-03 52.12481
#> 180  180 476.5843 4.023882e-03 447.42879 29.149102 2.341574e-03 52.41574
#> 181  181 476.2922 3.477192e-03 447.02683 29.259533 2.370779e-03 52.70779
#> 182  182 475.9991 3.003128e-03 446.62516 29.368506 2.400093e-03 53.00093
#> 183  183 475.7048 2.592293e-03 446.22379 29.476034 2.429515e-03 53.29515
#> 184  184 475.4096 2.236466e-03 445.82272 29.582134 2.459045e-03 53.59045
#> 185  185 475.1132 1.928464e-03 445.42197 29.686818 2.488679e-03 53.88679
#> 186  186 474.8158 1.662016e-03 445.02154 29.790102 2.518418e-03 54.18418
#> 187  187 474.5174 1.431647e-03 444.62143 29.892000 2.548259e-03 54.48259
#> 188  188 474.2180 1.232585e-03 444.22165 29.992526 2.578201e-03 54.78201
#> 189  189 473.9176 1.060672e-03 443.82220 30.091693 2.608244e-03 55.08244
#> 190  190 473.6162 9.122870e-04 443.42309 30.189516 2.638384e-03 55.38384
#> 191  191 473.3138 7.842792e-04 443.02432 30.286009 2.668622e-03 55.68622
#> 192  192 473.0104 6.739090e-04 442.62588 30.381184 2.698956e-03 55.98956
#> 193  193 472.7062 5.787972e-04 442.22780 30.475055 2.729384e-03 56.29384
#> 194  194 472.4009 4.968774e-04 441.83005 30.567636 2.759906e-03 56.59906
#> 195  195 472.0948 4.263560e-04 441.43265 30.658940 2.790519e-03 56.90519
#> 196  196 471.7878 3.656778e-04 441.03560 30.748979 2.821223e-03 57.21223
#> 197  197 471.4798 3.134951e-04 440.63890 30.837767 2.852016e-03 57.52016
#> 198  198 471.1710 2.686404e-04 440.24255 30.925316 2.882898e-03 57.82898
#> 199  199 470.8613 2.301034e-04 439.84655 31.011640 2.913867e-03 58.13867
#> 200  200 470.5508 1.970102e-04 439.45090 31.096749 2.944921e-03 58.44921
#> 201  201 470.2394 1.686052e-04 439.05560 31.180658 2.976060e-03 58.76060
#> 202  202 469.9272 1.442355e-04 438.66065 31.263378 3.007282e-03 59.07282
#> 203  203 469.6141 1.233375e-04 438.26606 31.344921 3.038586e-03 59.38586
#> 204  204 469.3003 1.054246e-04 437.87181 31.425300 3.069971e-03 59.69971
#> 205  205 468.9856 9.007739e-05 437.47792 31.504526 3.101436e-03 60.01436
#> 206  206 468.6702 7.693518e-05 437.08438 31.582610 3.132980e-03 60.32980
#> 207  207 468.3540 6.568565e-05 436.69119 31.659566 3.164601e-03 60.64601
#> 208  208 468.0370 5.605956e-05 436.29836 31.735404 3.196299e-03 60.96299
#> 209  209 467.7193 4.782614e-05 435.90587 31.810135 3.228072e-03 61.28072
#> 210  210 467.4008 4.078684e-05 435.51374 31.883772 3.259919e-03 61.59919
#> 211  211 467.0816 3.477091e-05 435.12196 31.956325 3.291839e-03 61.91839
#> 212  212 466.7617 2.963165e-05 434.73053 32.027805 3.323831e-03 62.23831
#> 213  213 466.4411 2.524304e-05 434.33945 32.098223 3.355894e-03 62.55894
#> 214  214 466.1197 2.149690e-05 433.94873 32.167591 3.388027e-03 62.88027
#> 215  215 465.7977 1.830040e-05 433.55835 32.235918 3.420229e-03 63.20229
#> 216  216 465.4750 1.557393e-05 433.16833 32.303216 3.452499e-03 63.52499
#> 217  217 465.1516 1.324924e-05 432.77866 32.369496 3.484835e-03 63.84835
#> 218  218 464.8276 1.126784e-05 432.38933 32.434767 3.517237e-03 64.17237
#> 219  219 464.5030 9.579663e-06 432.00036 32.499040 3.549704e-03 64.49704
#> 220  220 464.1777 8.141814e-06 431.61173 32.562325 3.582235e-03 64.82235
#> 221  221 463.8517 6.917608e-06 431.22346 32.624633 3.614829e-03 65.14829
#> 222  222 463.5252 5.875658e-06 430.83553 32.685973 3.647484e-03 65.47484
#> 223  223 463.1980 4.989132e-06 430.44796 32.746355 3.680200e-03 65.80200
#> 224  224 462.8702 4.235098e-06 430.06073 32.805790 3.712976e-03 66.12976
#> 225  225 462.5419 3.593967e-06 429.67385 32.864286 3.745811e-03 66.45811
#> 226  226 462.2130 3.049009e-06 429.28732 32.921854 3.778705e-03 66.78705
#> 227  227 461.8835 2.585945e-06 428.90113 32.978503 3.811655e-03 67.11655
#> 228  228 461.5534 2.192593e-06 428.51530 33.034243 3.844661e-03 67.44661
#> 229  229 461.2228 1.858561e-06 428.12981 33.089083 3.877723e-03 67.77723
#> 230  230 460.8916 1.574990e-06 427.74466 33.143032 3.910839e-03 68.10839